% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{addMissing}
\alias{addMissing}
\title{Add missing data}
\usage{
addMissing(data)
}
\arguments{
\item{data}{dataframe to test and convert}
}
\description{
Used to add missing data to input files and customize format
}
