% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argus.R
\docType{package}
\name{argus}
\alias{argus}
\title{argus}
\description{
\itemize{
\item Online App:
\item Github: https://github.com/JGCRI/argus
\item Webpage: https://jgcri.github.io/argus/
\item Cheatsheet: https://github.com/JGCRI/argus/blob/master/argusCheatsheet.pdf}
}
