% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{describe}
\usage{
describe(dat, screen_vars = c("username", "ip"), n_example_values = 5,
  max_nchar = 25, round_digits = 2, show_size = TRUE)
}
\arguments{
\item{dat}{dataframe or tibble}

\item{screen_vars}{vector of variable names whose values should not be show in the summary}

\item{n_example_values}{max number of value examples to show}

\item{max_nchar}{max nchar length of value examples}

\item{round_digits}{digits for rounding displayed numbers}

\item{show_size}{logical, whether to include the data size at the top of the summary}
}
\value{
a tibble
}
\description{
Make a summary dataframe for a dataframe.
}
\examples{
describe(tibble(x=c(1,2,3)))
}
