% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readgcam_functions.R
\name{elecInvest}
\alias{elecInvest}
\title{elecInvest}
\usage{
elecInvest(elec_gen_vintage, world_regions, start_year = 2010, end_year = 2050)
}
\arguments{
\item{elec_gen_vintage}{Electricity vintage query result}

\item{world_regions}{GCAM regions for which to collect data}

\item{start_year}{Start year of time frame of interest for analysis}

\item{end_year}{end_year of time frame of interest for analysis}
}
\value{
Returns data in a form required by readgcam.R
}
\description{
Function that calculates electricity subsector investment requirements from a given GCAM run.
}
\keyword{infrastructure}
\keyword{investments,}
