% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_GCAMv5p3_SSP235}
\alias{example_GCAMv5p3_SSP235}
\title{data
exampleData}
\format{
Tibble
}
\source{
Example run from GCAM v5.3
}
\usage{
example_GCAMv5p3_SSP235
}
\description{
data
exampleData
}
\examples{
\dontrun{
 library(argus);
 argus::example_GCAMv5p3_SSP235
}
}
\keyword{datasets}
