% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{exportHeight}
\alias{exportHeight}
\title{exportHeight}
\usage{
exportHeight(chartsperrow, max_height_in, numelement, lenperchart)
}
\arguments{
\item{chartsperrow}{number of columns}

\item{max_height_in}{max height}

\item{numelement}{number of planned elements}

\item{lenperchart}{height per element}
}
\description{
generate height for exported images
}
