% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{exportWidth}
\alias{exportWidth}
\title{exportWidth}
\usage{
exportWidth(max_width_in, numelement, lenperchart)
}
\arguments{
\item{max_width_in}{max width}

\item{numelement}{number of planned elements}

\item{lenperchart}{height per element}
}
\description{
generate width for exported images
}
