% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flag}
\alias{flag}
\title{flag}
\usage{
flag(df, level = 0.9)
}
\arguments{
\item{df}{tibble, as multiple rows produced by to_row()}

\item{level}{value between 0 and 1, confidence level used for prediction confidence interval}
}
\value{
a named list (names are variables from df, other than timestamp), each element is a list with components fit (prediction value), lwr, upr (lower and upper limits of prediction confindence interval), obs (observed value), flag (1 if obs>upr, -1 if obs<lwr, 0 otherwise).
}
\description{
Compare the latest in historical data with the prediction range from the earlier data
}
