% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid025}
\alias{grid025}
\title{grid025}
\format{
.csv
}
\source{
"inst/extdata/grid025.csv"
}
\usage{
grid025
}
\description{
grid025
}
\examples{
\dontrun{
 library(argus);
 argus::grid025
}
}
\keyword{datasets}
