% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid050}
\alias{grid050}
\title{grid050}
\format{
.csv
}
\source{
"inst/extdata/grid050.csv"
}
\usage{
grid050
}
\description{
grid050
}
\examples{
\dontrun{
 library(argus);
 argus::grid050
}
}
\keyword{datasets}
