% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdfFind.R
\name{mapdfFind}
\alias{mapdfFind}
\title{mapdfFind}
\usage{
mapdfFind(dataTbl)
}
\arguments{
\item{dataTbl}{Palette name to view the palette colors. Eg. colors("pal_Basic")}
}
\value{
dataframe with modified subRegions, subRegion shapefile and subRegion type
}
\description{
Given a data.frame with a subRegion column, this function searches for an appropriate map from the pre-loaded argus maps.
}
\keyword{find}
\keyword{map,}
