% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{maprs}
\alias{maprs}
\title{maprs}
\usage{
maprs(flag, mapLegend, mapYear, scenarioRefSelected, dataMapx, dataMapz)
}
\arguments{
\item{flag}{indicates which type of map plot to return; 1 for abs val, 2 for abs prcnt val, 3 for abs ref scen}

\item{mapLegend}{map legend text}

\item{mapYear}{year of map}

\item{scenarioRefSelected}{selected reference scenario}

\item{dataMapx}{dataMapx}

\item{dataMapz}{if flag is 2 or 3, use dataPrcntAbsMapx}
}
\description{
returns map of reference scenario
}
