% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{parse_local}
\alias{parse_local}
\title{parse_local}
\usage{
parse_local(inputdatapath, urlfiledatadatapath)
}
\arguments{
\item{inputdatapath}{local file to parse}

\item{urlfiledatadatapath}{local file to parse}
}
\description{
parse local files
}
