% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{plotAbs}
\alias{plotAbs}
\title{plotAbs}
\usage{
plotAbs(
  dataChartPlot,
  scenarioRefSelected,
  ggplottheme = ggplot2::theme_bw(),
  textsize = 15
)
}
\arguments{
\item{dataChartPlot}{dataChartPlot: dataChartx()}

\item{scenarioRefSelected}{scenarioRefSelected}

\item{ggplottheme}{ggplot theme to use}

\item{textsize}{Default = 15}
}
\description{
generate chart plot for absolute value
}
