% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{plotDiffAbs}
\alias{plotDiffAbs}
\title{plotDiff}
\usage{
plotDiffAbs(
  dataChartPlot,
  scenarioRefSelected,
  ggplottheme = ggplot2::theme_bw(),
  textsize = 15
)
}
\arguments{
\item{dataChartPlot}{dataChartPlot: dataDiffAbsx() or dataPrcntDiffx()}

\item{scenarioRefSelected}{scenarioRefSeleceted#'}

\item{ggplottheme}{ggplot theme to use}

\item{textsize}{Default = 15}
}
\description{
generate chart plot for absolute difference and percent difference
}
