% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{plotMap}
\alias{plotMap}
\title{plotMap}
\usage{
plotMap(
  mapData = NULL,
  legendType = "kmeans",
  mapX = NULL,
  scenRef = NULL,
  diff = NULL
)
}
\arguments{
\item{mapData}{mapData: dataMapx()}

\item{legendType}{Default = "kmeans". Legend type can be "pretty" or "kmeans".}

\item{mapX}{input$mapYear}

\item{scenRef}{scenRef}

\item{diff}{either "absolute", "percent" or NULL in the case of no diff}
}
\description{
generate chart plot for absolute value
}
