% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preloaded_data.R
\name{preloaded_data}
\alias{preloaded_data}
\title{mappings}
\usage{
preloaded_data()
}
\value{
A dataframe with preloaded data
}
\description{
preloaded_data that allows Argus to load up data from a github or zenodo link.
Argus data is maintained on the https://github.com/JGCRI/argusdata repo
If new data is added to the repo users can directly change this file and
commit on github at : https://github.com/JGCRI/argus/blob/main/R/preloaded_data.R
}
\examples{
library(rmap)
preloaded_data()
}
\keyword{data}
\keyword{preloaded}
