% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataviz.R
\docType{package}
\name{rdataviz}
\alias{rdataviz}
\title{rdataviz}
\description{
\itemize{
\item Online App: 
\item Github: https://github.com/JGCRI/rdataviz
\item Webpage: https://jgcri.github.io/rdataviz/
\item Cheatsheet: https://github.com/JGCRI/rdataviz/blob/master/rmapCheatsheet.pdf}
}
