% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{summaryPlot}
\alias{summaryPlot}
\title{summaryPlot}
\usage{
summaryPlot(aspectratio, textsize, dataSumx, ggplottheme = ggplot2::theme_bw())
}
\arguments{
\item{aspectratio}{aspect ratio}

\item{textsize}{text size}

\item{dataSumx}{dataSumx()}

\item{ggplottheme}{ggplot theme to use}
}
\description{
generate summary plot
}
