% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{summaryPlotReg}
\alias{summaryPlotReg}
\title{summaryPlotReg}
\usage{
summaryPlotReg(
  dataMapx,
  ggplottheme = ggplot2::theme_bw(),
  textsize = 15,
  subsetRegionsx
)
}
\arguments{
\item{dataMapx}{Input map dataframe}

\item{ggplottheme}{ggplot theme to use}

\item{textsize}{Default =15}

\item{subsetRegionsx}{Subset of regions}
}
\description{
generate summary plot
}
