% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{to_row}
\alias{to_row}
\title{to_row}
\usage{
to_row(df, timestamp = NULL)
}
\arguments{
\item{df}{- table produced by describe()}

\item{timestamp}{- if not given, will be Sys.time()}
}
\value{
a tibble
}
\description{
Convert the summary table (as produced by describe()) to a 1-row wide format, adding the timestamp column
}
\examples{
to_row(describe(tibble(x=c(1,2,3))))
}
