% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xmlQueries}
\alias{xmlQueries}
\title{xmlQueries xml file}
\format{
.xml
}
\source{
rdataviz
}
\usage{
xmlQueries
}
\description{
xmlQueries xml file
}
\examples{
\dontrun{
 library(rdataviz); library(XML)
 rdataviz::xmlQueries
 # Can save xml
 XML::saveXML(rdataviz::xmlQueries, file=paste(getwd(), "/xmlQueries.xml", sep = ""))
}
}
\keyword{datasets}
