% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{breaks}
\alias{breaks}
\title{breaks}
\usage{
breaks(dataMap_raw_param = NULL, breaks_n = 7)
}
\arguments{
\item{dataMap_raw_param}{Input dataframe}

\item{breaks_n}{Number of breaks}
}
\description{
returns breaks generated using k-mean and pretty
}
