% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\title{constants}
\usage{
constants(name = NULL)
}
\arguments{
\item{name}{Default=NULL. Name of constant to call.}
}
\value{
A list of constants
}
\description{
This function holds the various constants used in the rdataviz package
}
\details{
List of Constants
\itemize{
\item "chosenMix"}
}
\examples{
\dontrun{
library(rdataviz)
constants()$chosenMix
}
}
\keyword{constants}
