% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{plotAbs}
\alias{plotAbs}
\title{plotAbs}
\usage{
plotAbs(dataChartPlot, scenarioRefSelected)
}
\arguments{
\item{dataChartPlot}{dataChartPlot: dataChartx()}

\item{scenarioRefSelected}{scenarioRefSelected}
}
\description{
generate chart plot for absolute value
}
