% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{plotDiff}
\alias{plotDiff}
\title{plotDiff}
\usage{
plotDiff(dataChartPlot, scenarioRefSelected)
}
\arguments{
\item{dataChartPlot}{dataChartPlot: dataDiffAbsx() or dataPrcntDiffx()}

\item{scenarioRefSelected}{scenarioRefSeleceted}
}
\description{
generate chart plot for absolute difference and percent difference
}
