% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{process_map}
\alias{process_map}
\title{process map}
\usage{
process_map(dataMap_raw, i, mapLegend, mapYear)
}
\arguments{
\item{dataMap_raw}{input dataframe}

\item{i}{parameter/category for dataframe}
}
\description{
returns map
}
