% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{summaryPlot}
\alias{summaryPlot}
\title{summaryPlot}
\usage{
summaryPlot(aspectratio, textsize, titletext, dataSumx)
}
\arguments{
\item{aspectratio}{aspect ratio}

\item{textsize}{text size}

\item{titletext}{title text}

\item{dataSumx}{dataSumx}
}
\description{
generate summary plot
}
